huge, huge update to work with the new utils API
Updating to work with the new utils API allows me to use databases to
read data for processing. There are a few other important changes
including big updates to the EntityRMRunner and DoubleEntityRMRunner
classes, and those related to them. Again, these are intended to be more
flexible with the primary purpose of allowing me to use database data
sources.